These progress bar changes are a WIP and currently broken
The progress bar changes here attempt to nest the images so that the background
of the png helmets are auto set to the current drawing of the form.
This simulates transparency, but I have been getting BSoDs which is obviously
not ideal... Changes in this direction need to follow this path and avoid the
BSoD cause or instead some drawing functions need to be overloaded...

This commit and branch is just to save my progress for later review, but
tbh I don't have the time to invest this much effort into making image
transparency work in forms. Its a dumb problem. Maybe I will get around
to it soon(TM).